remove handlers module and move lms and lti existing handlers to its own and specific module
