+++ ansible/roles/jupyterhub/files/jupyterhub_config_lti11.py
@@ -4,6 +4,7 @@ import sys
 
 from illumidesk.authenticators.authenticator import LTI11Authenticator
 from illumidesk.authenticators.authenticator import setup_course_hook
+from illumidesk.grades.handlers import SendGradesHandler
 from illumidesk.spawners.spawner import IllumiDeskDockerSpawner
 
 
@@ -215,7 +216,7 @@ c.DockerSpawner.name_template = 'jupyter-{raw_username}'
 c.JupyterHub.extra_handlers = [
     (
         r'/submit-grades/(?P<course_id>[a-zA-Z0-9-_]+)/(?P<assignment_name>.*)$',
-        'illumidesk.handlers.lms_grades.SendGradesHandler',
+        SendGradesHandler,
     ),
 ]
